+++ client/navigation/index.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
-import { useAuthentication } from '../utils/hooks/useAuth';
-import UserStack from './authStack';
-import AuthStack from './userStack';
+// import React from 'react';
+// import { useAuthentication } from '../utils/hooks/useAuth';
+// import UserStack from './authStack';
+// import AuthStack from './userStack';
 
-export default function RootNavigation() {
-  const { user } = useAuthentication();
+// export default function RootNavigation() {
+//   const { user } = useAuthentication();
 
-  return user ? <UserStack /> : <AuthStack />;
-}
+//   return user ? <UserStack /> : <AuthStack />;
+// }
